pyproject.toml
[tool.poetry]
name = "demo-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "demo_app"}]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.1"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
line-bot-sdk = "^2.3.0"
python-dotenv = "^0.21.0"
aiolinebot = "^0.4.1"
sqlalchemy = "^1.4.42"
aiomysql = "^0.1.1"
python-multipart = "^0.0.5"
pillow = "^9.2.0"
jinja2 = "^3.1.2"
cloudinary = "^1.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
